+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 - Node support Meta class
 - Add stale option to settings/cacheme decorator/node meta
 - Node support hit/miss function
+- Cacheme tags using nodes instead of cacheme instances
+
+### Removed
+- Remove `invalid_all()` method from tag, using `tag.objects.invalid()` instead
 
 ## [0.1.0]
 ### Added